update membership scheme
