+++ index.ts
@@ -8,12 +8,10 @@ export * from "./src/base/event/event";
 export * from "./src/base/event/sourceCodeLine";
 
 export * from "./src/dbScheme/businessOperation";
-export * from "./src/dbScheme/confirmedMember";
 export * from "./src/dbScheme/groupedEvent";
-export * from "./src/dbScheme/membership";
 export * from "./src/dbScheme/notificationsChannels";
 export * from "./src/dbScheme/notificationsChannelSettings";
-export * from "./src/dbScheme/pendingMember";
+export * from "./src/dbScheme/membership";
 export * from "./src/dbScheme/plan";
 export * from "./src/dbScheme/project";
 export * from "./src/dbScheme/projectNotificationsRule";
