--- src/dbScheme/pendingMember.ts
@@ -1,17 +1,0 @@
-import { ObjectId } from "mongodb";
-
-
-/**
- * Represents pending member info in DB
- */
-export interface PendingMemberDBScheme {
-    /**
-     * Document id
-     */
-    _id: ObjectId;
-
-    /**
-     * User email for invitation
-     */
-    userEmail: string;
-}
